Split and clean
